+++ R/mod_tab_curated_analytes.R
@@ -224,14 +224,7 @@ mod_tab_curated_analytes_server <- function(id, info, cluster, biogroup_column){
       
       return(to_return)
     })
-    
-    
-    # Remove "Curating analytes..." spinner
-    observe({
-      req(curated_analytes_table())
-      shinybusy::remove_modal_spinner()
-    })
-
+  
     
     return(list(plot = curated_analytes_plot,
                 analytes_to_include = analytes_to_include))